Added logic for getting claim details + files + some UI rework
