+++ Projects/App/Sources/Service/OctopusClientsImplementation/FetchClaimDetailsClientOctopus.swift
@@ -8,13 +8,38 @@ class FetchClaimDetailsClientOctopus: hFetchClaimDetailsClient {
 
     func get(for type: FetchClaimDetailsType) async throws -> ClaimModel {
         switch type {
-        case let .claim(id):
-            let query = OctopusGraphQL.ClaimsQuery()
-            let data = try await octopus.client.fetch(query: query, cachePolicy: .fetchIgnoringCacheCompletely)
-            if let claimFragment = data.currentMember.claims.first(where: { $0.id == id })?.fragments.claimFragment {
-                return ClaimModel(claim: claimFragment)
+        case let .claim(id, claimStatus):
+            switch claimStatus {
+            case .active:
+                if Dependencies.featureFlags().isClaimHistoryEnabled {
+                    let query = OctopusGraphQL.ActiveClaimsQuery()
+                    let data = try await octopus.client.fetch(query: query, cachePolicy: .fetchIgnoringCacheCompletely)
+                    if let claimFragment = data.currentMember.claimsActive.first(where: { $0.id == id })?.fragments
+                        .claimFragment
+                    {
+                        return ClaimModel(claim: claimFragment)
+                    }
+                    throw FetchClaimDetailsError.noClaimFound
+                } else {
+                    let query = OctopusGraphQL.ClaimsQuery()
+                    let data = try await octopus.client.fetch(query: query, cachePolicy: .fetchIgnoringCacheCompletely)
+                    if let claimFragment = data.currentMember.claims.first(where: { $0.id == id })?.fragments
+                        .claimFragment
+                    {
+                        return ClaimModel(claim: claimFragment)
+                    }
+                    throw FetchClaimDetailsError.noClaimFound
+                }
+            case .history:
+                let query = OctopusGraphQL.HistoryClaimsQuery()
+                let data = try await octopus.client.fetch(query: query, cachePolicy: .fetchIgnoringCacheCompletely)
+                if let claimFragment = data.currentMember.claimsHistory.first(where: { $0.id == id })?.fragments
+                    .claimFragment
+                {
+                    return ClaimModel(claim: claimFragment)
+                }
+                throw FetchClaimDetailsError.noClaimFound
             }
-            throw FetchClaimDetailsError.noClaimFound
         case let .conversation(id):
             let query = OctopusGraphQL.ClaimFromConversationQuery(conversationId: id)
             let data = try await octopus.client.fetch(query: query, cachePolicy: .fetchIgnoringCacheCompletely)
@@ -25,17 +50,41 @@ class FetchClaimDetailsClientOctopus: hFetchClaimDetailsClient {
         }
     }
 
-    func getFiles(for type: FetchClaimDetailsType) async throws -> (claimId: String, files: [hCore.File]) {
+    func getFiles(for type: FetchClaimDetailsType) async throws -> (claimId: String, files: [File]) {
         switch type {
-        case let .claim(id):
-            let query = OctopusGraphQL.ClaimsFilesQuery()
-            let data = try await octopus.client.fetch(query: query, cachePolicy: .fetchIgnoringCacheCompletely)
+        case let .claim(id, claimStatus):
+            switch claimStatus {
+            case .active:
+                if Dependencies.featureFlags().isClaimHistoryEnabled {
+                    let query = OctopusGraphQL.ActiveClaimsFilesQuery()
+                    let data = try await octopus.client.fetch(query: query, cachePolicy: .fetchIgnoringCacheCompletely)
 
-            if let files = data.currentMember.claims.first(where: { $0.id == id })?.files
-                .compactMap({ File(with: $0.fragments.fileFragment) })
-            {
-                return (id, files)
+                    if let files = data.currentMember.claimsActive.first(where: { $0.id == id })?.files
+                        .compactMap({ File(with: $0.fragments.fileFragment) })
+                    {
+                        return (id, files)
+                    }
+                } else {
+                    let query = OctopusGraphQL.ClaimsFilesQuery()
+                    let data = try await octopus.client.fetch(query: query, cachePolicy: .fetchIgnoringCacheCompletely)
+
+                    if let files = data.currentMember.claims.first(where: { $0.id == id })?.files
+                        .compactMap({ File(with: $0.fragments.fileFragment) })
+                    {
+                        return (id, files)
+                    }
+                }
+            case .history:
+                let query = OctopusGraphQL.HistoryClaimsFilesQuery()
+                let data = try await octopus.client.fetch(query: query, cachePolicy: .fetchIgnoringCacheCompletely)
+
+                if let files = data.currentMember.claimsHistory.first(where: { $0.id == id })?.files
+                    .compactMap({ File(with: $0.fragments.fileFragment) })
+                {
+                    return (id, files)
+                }
             }
+
             throw FetchClaimDetailsError.noClaimFound
         case let .conversation(id):
             let query = OctopusGraphQL.ClaimFilesFromConversationQuery(conversationId: id)
